// generated/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Balance = {
  __typename?: 'Balance';
  /** Coin type for the balance, such as `0x2::sui::SUI`. */
  coinType?: Maybe<MoveType>;
  /** The total balance across all coin objects of this coin type. */
  totalBalance?: Maybe<Scalars['BigInt']['output']>;
};
// ...
export type CoinMetadata = IAddressable & IMoveObject & IObject & {
// ...
  objects?: Maybe<MoveObjectConnection>;
// ...
};
// ...
export type Display = {
  __typename?: 'Display';
  /** If any fields failed to render, this will contain a mapping from failed field names to error messages. If all fields succeed, this will be `null`. */
  errors?: Maybe<Scalars['JSON']['output']>;
  /** Output for all successfully substituted display fields. Unsuccessful fields will be `null`, and will be accompanied by a field in `errors`, explaining the error. */
  output?: Maybe<Scalars['JSON']['output']>;
};
// ...
export type MoveObject = IAddressable & IMoveObject & IObject & {
  __typename?: 'MoveObject';
  /** The MoveObject's ID. */
  address: Scalars['SuiAddress']['output'];
  /** Attempts to convert the object into a CoinMetadata. */
  asCoinMetadata?: Maybe<CoinMetadata>;
  /** Attempts to convert the object into a DynamicField. */
  asDynamicField?: Maybe<DynamicField>;
  /**
   * Fetch the total balance for coins with marker type `coinType` (e.g. `0x2::sui::SUI`), owned by this address.
   *
   * If the address does not own any coins of that type, a balance of zero is returned.
   */
  balance?: Maybe<Balance>;
  /** Total balance across coins owned by this address, grouped by coin type. */
  balances?: Maybe<BalanceConnection>;
  /** The structured representation of the object's contents. */
  contents?: Maybe<MoveValue>;
  /** The domain explicitly configured as the default SuiNS name for this address. */
  defaultSuinsName?: Maybe<Scalars['String']['output']>;
  /** 32-byte hash that identifies the object's contents, encoded in Base58. */
  digest?: Maybe<Scalars['String']['output']>;
  /**
   * Access a dynamic field on an object using its type and BCS-encoded name.
   *
   * Returns `null` if a dynamic field with that name could not be found attached to this object.
   */
  dynamicField?: Maybe<DynamicField>;
  /**
   * Dynamic fields owned by this object.
   *
   * Dynamic fields on wrapped objects can be accessed using `Address.dynamicFields`.
   */
  dynamicFields?: Maybe<DynamicFieldConnection>;
  /**
   * Access a dynamic object field on an object using its type and BCS-encoded name.
   *
   * Returns `null` if a dynamic object field with that name could not be found attached to this object.
   */
  dynamicObjectField?: Maybe<DynamicField>;
  /**
   * Whether this object can be transfered using the `TransferObjects` Programmable Transaction Command or `sui::transfer::public_transfer`.
   *
   * Both these operations require the object to have both the `key` and `store` abilities.
   */
  hasPublicTransfer?: Maybe<Scalars['Boolean']['output']>;
  /** The Base64-encoded BCS serialize of this object, as a `MoveObject`. */
  moveObjectBcs?: Maybe<Scalars['Base64']['output']>;
  /**
   * Fetch the total balances keyed by coin types (e.g. `0x2::sui::SUI`) owned by this address.
   *
   * If the address does not own any coins of a given type, a balance of zero is returned for that type.
   */
  multiGetBalances?: Maybe<Array<Balance>>;
  /**
   * Access dynamic fields on an object using their types and BCS-encoded names.
   *
   * Returns a list of dynamic fields that is guaranteed to be the same length as `keys`. If a dynamic field in `keys` could not be found in the store, its corresponding entry in the result will be `null`.
   */
  multiGetDynamicFields: Array<Maybe<DynamicField>>;
  /**
   * Access dynamic object fields on an object using their types and BCS-encoded names.
   *
   * Returns a list of dynamic object fields that is guaranteed to be the same length as `keys`. If a dynamic object field in `keys` could not be found in the store, its corresponding entry in the result will be `null`.
   */
  multiGetDynamicObjectFields: Array<Maybe<DynamicField>>;
  /**
   * Fetch the object with the same ID, at a different version, root version bound, or checkpoint.
   *
   * If no additional bound is provided, the latest version of this object is fetched at the latest checkpoint.
   */
  objectAt?: Maybe<Object>;
  /** The Base64-encoded BCS serialization of this object, as an `Object`. */
  objectBcs?: Maybe<Scalars['Base64']['output']>;
  /** Paginate all versions of this object after this one. */
  objectVersionsAfter?: Maybe<ObjectConnection>;
  /** Paginate all versions of this object before this one. */
  objectVersionsBefore?: Maybe<ObjectConnection>;
  /** Objects owned by this object, optionally filtered by type. */
  objects?: Maybe<MoveObjectConnection>;
  /** The object's owner kind. */
  owner?: Maybe<Owner>;
  /** The transaction that created this version of the object. */
  previousTransaction?: Maybe<Transaction>;
  /** The transactions that sent objects to this object. */
  receivedTransactions?: Maybe<TransactionConnection>;
  /** The SUI returned to the sponsor or sender of the transaction that modifies or deletes this object. */
  storageRebate?: Maybe<Scalars['BigInt']['output']>;
  /** The version of this object that this content comes from. */
  version?: Maybe<Scalars['UInt53']['output']>;
};
// ...
export type MoveValue = {
  __typename?: 'MoveValue';
  /** The BCS representation of this value, Base64-encoded. */
  bcs?: Maybe<Scalars['Base64']['output']>;
  /**
   * A rendered JSON blob based on an on-chain template, substituted with data from this value.
   *
   * Returns `null` if the value's type does not have an associated `Display` template.
   */
  display?: Maybe<Display>;
  /**
   * Representation of a Move value in JSON, where:
   *
   * - Addresses, IDs, and UIDs are represented in canonical form, as JSON strings.
   * - Bools are represented by JSON boolean literals.
   * - u8, u16, and u32 are represented as JSON numbers.
   * - u64, u128, and u256 are represented as JSON strings.
   * - Balances, Strings, and Urls are represented as JSON strings.
   * - Vectors of bytes are represented as Base64 blobs, and other vectors are represented by JSON arrays.
   * - Structs are represented by JSON objects.
   * - Enums are represented by JSON objects, with a field named `@variant` containing the variant name.
   * - Empty optional values are represented by `null`.
   */
  json?: Maybe<Scalars['JSON']['output']>;
  /** The value's type. */
  type?: Maybe<MoveType>;
};
// ...
export type Object = IAddressable & IObject & {
  __typename?: 'Object';
  /** The Object's ID. */
  address: Scalars['SuiAddress']['output'];
  /** Attempts to convert the object into a MoveObject. */
  asMoveObject?: Maybe<MoveObject>;
  /** Attempts to convert the object into a MovePackage. */
  asMovePackage?: Maybe<MovePackage>;
  /**
   * Fetch the total balance for coins with marker type `coinType` (e.g. `0x2::sui::SUI`), owned by this address.
   *
   * If the address does not own any coins of that type, a balance of zero is returned.
   */
  balance?: Maybe<Balance>;
  /** Total balance across coins owned by this address, grouped by coin type. */
  balances?: Maybe<BalanceConnection>;
  /** The domain explicitly configured as the default SuiNS name for this address. */
  defaultSuinsName?: Maybe<Scalars['String']['output']>;
  /** 32-byte hash that identifies the object's contents, encoded in Base58. */
  digest?: Maybe<Scalars['String']['output']>;
  /**
   * Access a dynamic field on an object using its type and BCS-encoded name.
   *
   * Returns `null` if a dynamic field with that name could not be found attached to this object.
   */
  dynamicField?: Maybe<DynamicField>;
  /** Dynamic fields owned by this object. */
  dynamicFields?: Maybe<DynamicFieldConnection>;
  /**
   * Access a dynamic object field on an object using its type and BCS-encoded name.
   *
   * Returns `null` if a dynamic object field with that name could not be found attached to this object.
   */
  dynamicObjectField?: Maybe<DynamicField>;
  /**
   * Fetch the total balances keyed by coin types (e.g. `0x2::sui::SUI`) owned by this address.
   *
   * Returns `None` when no checkpoint is set in scope (e.g. execution scope).
   * If the address does not own any coins of a given type, a balance of zero is returned for that type.
   */
  multiGetBalances?: Maybe<Array<Balance>>;
  /**
   * Access dynamic fields on an object using their types and BCS-encoded names.
   *
   * Returns a list of dynamic fields that is guaranteed to be the same length as `keys`. If a dynamic field in `keys` could not be found in the store, its corresponding entry in the result will be `null`.
   */
  multiGetDynamicFields: Array<Maybe<DynamicField>>;
  /**
   * Access dynamic object fields on an object using their types and BCS-encoded names.
   *
   * Returns a list of dynamic object fields that is guaranteed to be the same length as `keys`. If a dynamic object field in `keys` could not be found in the store, its corresponding entry in the result will be `null`.
   */
  multiGetDynamicObjectFields: Array<Maybe<DynamicField>>;
  /**
   * Fetch the object with the same ID, at a different version, root version bound, or checkpoint.
   *
   * If no additional bound is provided, the latest version of this object is fetched at the latest checkpoint.
   */
  objectAt?: Maybe<Object>;
  /** The Base64-encoded BCS serialization of this object, as an `Object`. */
  objectBcs?: Maybe<Scalars['Base64']['output']>;
  /** Paginate all versions of this object after this one. */
  objectVersionsAfter?: Maybe<ObjectConnection>;
  /** Paginate all versions of this object before this one. */
  objectVersionsBefore?: Maybe<ObjectConnection>;
  /** Objects owned by this object, optionally filtered by type. */
  objects?: Maybe<MoveObjectConnection>;
  /** The object's owner kind. */
  owner?: Maybe<Owner>;
  /** The transaction that created this version of the object. */
  previousTransaction?: Maybe<Transaction>;
  /** The transactions that sent objects to this object */
  receivedTransactions?: Maybe<TransactionConnection>;
  /** The SUI returned to the sponsor or sender of the transaction that modifies or deletes this object. */
  storageRebate?: Maybe<Scalars['BigInt']['output']>;
  /** The version of this object that this content comes from. */
  version?: Maybe<Scalars['UInt53']['output']>;
};
// ...
export type ObjectConnection = {
  __typename?: 'ObjectConnection';
  /** A list of edges. */
  edges: Array<ObjectEdge>;
  /** A list of nodes. */
  nodes: Array<Object>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};

/** An edge in a connection. */
export type ObjectEdge = {
  __typename?: 'ObjectEdge';
  /** A cursor for use in pagination */
  cursor: Scalars['String']['output'];
  /** The item at the end of the edge */
  node: Object;
};
// ...
export type Owner = AddressOwner | ConsensusAddressOwner | Immutable | ObjectOwner | Shared;
// ...
export type PageInfo = {
  __typename?: 'PageInfo';
  /** When paginating forwards, the cursor to continue. */
  endCursor?: Maybe<Scalars['String']['output']>;
  /** When paginating forwards, are there more items? */
  hasNextPage: Scalars['Boolean']['output'];
  /** When paginating backwards, are there more items? */
  hasPreviousPage: Scalars['Boolean']['output'];
  /** When paginating backwards, the cursor to continue. */
  startCursor?: Maybe<Scalars['String']['output']>;
};
// ...
export type Query = {
// ...
  objects?: Maybe<ObjectConnection>;
// ...
};
// ...
export type GetNftsByTypeQuery = { __typename?: 'Query', objects?: { __typename?: 'ObjectConnection', pageInfo: { __typename?: 'PageInfo', hasNextPage: boolean, endCursor?: string | null }, nodes: Array<{ __typename?: 'Object', address: any, asMoveObject?: { __typename?: 'MoveObject', contents?: { __typename?: 'MoveValue', json?: any | null, display?: { __typename?: 'Display', output?: any | null } | null } | null } | null }> } | null };